Refactroring v0.2

+++ ElectricityMetering.Core/Model/BalanceRate.cs
@@ -9,8 +9,8 @@ namespace ElectricityMetering.Core.Model
     public class BalanceRate
     {
         public int Id { get; set; }
-        public double Debt { get; set; }
-        public double Advance { get; set; }
-        public double Balance { get; set; }
+        public decimal Debt { get; set; }
+        public decimal Advance { get; set; }
+        public decimal Balance { get; set; }
     }
 }
