+++ ElectricityMetering.Core/Model/Payment.cs
@@ -10,9 +10,9 @@ namespace ElectricityMetering.Core.Model
     {
         public int Id { get; set; }
         public DateOnly Date { get; set; }
-        public double Cash { get; set; }
-        public double NonCash { get; set; }
-        public double Total { get; set; }
+        public decimal Cash { get; set; }
+        public decimal NonCash { get; set; }
+        public decimal Total { get; set; }
         public Owner Owner { get; set; } = null!;
     }
 }
